+++ scripts/generate-sample-graphs/generate-sample-graphs.ts
@@ -90,6 +90,7 @@ async function generateSampleGraphs({ isCommited }: { isCommited: boolean }) {
                           direction TB
 
                           input((Input)):::input
+                          %% [🌅] + Add it into README
                           ${block(
                               promptbookJson.promptTemplates
                                   .flatMap(({ title, dependentParameterNames, resultingParameterName }) => [
@@ -152,6 +153,6 @@ async function generateSampleGraphs({ isCommited }: { isCommited: boolean }) {
 }
 
 /**
- * TODO: !!!! Explicit output parameters
+ * TODO: [🌅] Explicit output parameters
  * TODO: Maybe use some Mermaid library instead of string templating
  */