+++ CHANGELOG.md
@@ -127,6 +127,14 @@ Cleanup and renaming
 -   Change command `PTBK_NAME` to `PROMPTBOOK_NAME` _(but keep backward compatibility and preserve alias `PTBK`)_
 -   Rename `runRemoteServer` -> `startRemoteServer` and return `Destroyable` object
 
+## 0.37.0 _(2024-03-07)_
+
+Explicit output parameters
+
+-   Every promptbook has to have `OUTPUT PARAMETER` property in header
+
+---
+
 ## Upcomming
 
 -   Rename "natural" -> "llm"